backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pytest
pytest-asyncio
httpx
python-dotenv